Update SociogramPrompts to new spec
