+++ src/components/Prompts/index.js
@@ -1,1 +1,2 @@
 export { default } from './Prompts';
+export { default as withSubjectNodeType } from './withSubjectNodeType';
